# src/main/resources/openapi.yml
openapi: 3.0.0
info:
    title: nhatrotot's backend
    version: 1.0.0
servers:
    - url: http://localhost:8080/api
...
paths:
...
    /auth/users/{userId}:
        delete:
            tags:
                - "auth"
            operationId: banUserHandle
            description: sub-admin or admin toggle activating user
            parameters:
                - name: userId
                  in: path
                  description: id of the user
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                200:
                    description: user delete successfully
                400:
                    $ref: "#/components/responses/BadRequest"
                403:
                    description: User need higher role to do this thing
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BadRequestResponse"
    /users:
        get:
            tags:
                - "user"
            operationId: getListUserInformation
            description: get list of users including user's public information
            parameters:
                - name: page
                  in: query
                  description: Page number
                  schema:
                      type: integer
                      minimum: 0
                      default: 0
                  required: false
                - name: size
                  in: query
                  schema:
                      type: integer
                      minimum: 1
                      default: 20
                  description: Page size
                  required: false
            responses:
                200:
                    description: list of user info and paging details
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserPublicInformationPage"
                400:
                    $ref: "#/components/responses/BadRequest"
    /users/{userId}:
        get:
            tags:
                - "user"
            operationId: getUserInformation
            description: return information based on user's id
            parameters:
                - name: userId
                  in: path
                  description: id of the user
                  schema:
                      type: string
                      format: uuid
                  required: true
            responses:
                200:
                    description: user's public information
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserPublicInformation"
                400:
                    $ref: "#/components/responses/BadRequest"
                404:
                    description: User not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BadRequestResponse"
    /users/me/my-info:
        get:
            tags:
                - "user"
            operationId: getMyInformation
            description: return user's information
            responses:
                200:
                    description: user's information
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserInformation"
                400:
                    $ref: "#/components/responses/BadRequest"
                404:
                    description: User not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BadRequestResponse"
        put:
            tags:
                - "user"
            operationId: changeUserInformationHandler
            description: change user information
            requestBody:
                required: true
                description: user changeable fields
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserFields"
            responses:
                200:
                    description: ok
                400:
                    $ref: "#/components/responses/BadRequest"
    /users/admin/users:
        get:
            tags:
                - "user"
            operationId: adminGetListUserInformation
            description: get list of users including user's public information
            parameters:
                - name: page
                  in: query
                  description: Page number
                  schema:
                      type: integer
                      minimum: 0
                      default: 0
                  required: false
                - name: size
                  in: query
                  schema:
                      type: integer
                      minimum: 1
                      default: 20
                  description: Page size
                  required: false
                - name: email
                  in: query
                  schema:
                      type: string
                      example: yuh
                      default: ""
                  description: email of user to find
                  required: false
                - name: role
                  in: query
                  schema:
                      type: integer
                      minimum: 1
                      maximum: 3
                      format: int64
                  description: role id of user to find
                  required: false
            responses:
                200:
                    description: list of user info and paging details
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserInformationPage"
                400:
                    $ref: "#/components/responses/BadRequest"
    /users/admin/users/{userId}:
        get:
            tags:
                - "user"
            operationId: adminGetUserInformation
            description: return information based on user's id
            parameters:
                - name: userId
                  in: path
                  description: id of the user
                  schema:
                      type: string
                      format: uuid
                  required: true
            responses:
                200:
                    description: user's information
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserInformation"
                400:
                    $ref: "#/components/responses/BadRequest"
                404:
                    description: User not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BadRequestResponse"
    /users/admin/users/{userId}/title:
        put:
            tags:
                - "user"
            operationId: changeUserTitleHandle
            description: change user's title
            parameters:
                - name: userId
                  in: path
                  description: id of the user
                  schema:
                      type: string
                      format: uuid
                  required: true
            requestBody:
                required: true
                description: user's email
                content:
                    application/json:
                        schema:
                            properties:
                                title:
                                    title: user's title
                                    type: integer
                                    minimum: 2
                                    maximum: 3
                                    format: int64
                            required:
                                - title
            responses:
                200:
                    description: ok
                400:
                    $ref: "#/components/responses/BadRequest"
                404:
                    description: User not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BadRequestResponse"
    /users/action/rating:
        post:
            tags:
                - "user"
            operationId: addRatingHandle
            description: add user rating
            requestBody:
                required: true
                description: user's email
                content:
                    application/json:
                        schema:
                            properties:
                                rating:
                                    title: user's rating
                                    type: integer
                                    minimum: 1
                                    maximum: 10
                                userId:
                                    type: string
                                    format: uuid
                            required:
                                - title
                                - userId
            responses:
                200:
                    description: ok
                400:
                    $ref: "#/components/responses/BadRequest"
                404:
                    description: User not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BadRequestResponse"
components:
    responses:
        BadRequest:
            description: Bad request.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/BadRequestResponse"
    schemas:
...
        BadRequestResponse:
            title: bad request.
            description: there is an error happening when server handle request.
            type: object
            properties:
                className:
                    type: string
                    title: name of the error
                    description: class name of the error
                    example: FileMaximumExceedException
                exMessage:
                    type: string
                    title: error message
                    description: error message
                    example: File not valid
...
        UserFields:
            title: user fields
            description: Editable fields of a user type.
            type: object
            properties:
                gender:
                    type: string
                    title: user's gender
                    enum:
                        - male
                        - female
                        - other
                phoneNumber:
                    type: string
                    title: user's phone number
                    pattern: '^\d{1,12}$'
                intro:
                    type: string
                    title: user's intro
                image:
                    type: string
                    title: user's image url
                birthDate:
                    type: string
                    title: user's birth date
                    format: date
                    example: 2017-07-21
            required:
                - gender
                - phoneNumber
                - intro
                - image
                - birthDate
        PublicUserField:
            title: user fields that are public fields
            allOf:
                - $ref: "#/components/schemas/UserFields"
                - type: object
                  properties:
                      firstName:
                          title: user's first name
                          type: string
                      lastName:
                          title: user's last name
                          type: string
                      email:
                          title: user's email
                          pattern: '^[^\s@]+@[^\s@]+\.[^\s@]+$'
                          type: string
                      registerAt:
                          title: user's created Date
                          type: string
                          format: date-time
                          example: 2017-07-21T17:32:28Z
                      banned:
                          title: user's status
                          type: boolean
                          nullable: true
                      avgRating:
                          title: user's average rating
                          type: number
            required:
                - firstName
                - lastName
                - email
                - registerAt
...
        UserPublicInformation:
            title: user public information
            description: omit balance
            allOf:
                - $ref: "#/components/schemas/PublicUserField"
                - type: object
                  properties:
                      title:
                          type: string
                      role:
                          type: string
                      userId:
                          type: string
                          format: uuid
        UserPublicInformationPage:
            title: user public information page
            description: user info page including total page and current page
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: "#/components/schemas/UserPublicInformation"
                totalPage:
                    type: integer
                    example: 10
                currentPage:
                    type: integer
                    example: 0
        UserInformation:
            title: user information
            description: balance
            allOf:
                - $ref: "#/components/schemas/UserPublicInformation"
                - type: object
                  properties:
                      balance:
                          type: number
        UserInformationPage:
            title: user public information page
            description: user info page including total page and current page
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: "#/components/schemas/UserInformation"
                totalPage:
                    type: integer
                    example: 10
                currentPage:
                    type: integer
                    example: 0
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT # optional, for documentation purposes only
security:
    - bearerAuth: []